+++ src/Nox.Generator/Nox.Generator.cs
@@ -5,7 +5,6 @@ using Nox.Generator.Application.EventGenerator;
 using Nox.Generator.Common;
 using Nox.Generator.Domain.CqrsGenerators;
 using Nox.Generator.Domain.DomainEventGenerator;
-using Nox.Generator.Presentation.Rest;
 using Nox.Solution;
 using System;
 using System.Collections.Generic;
@@ -17,6 +16,7 @@ using Nox.Generator.Infrastructure.Persistence.DbContextGenerator;
 using YamlDotNet.Core;
 using YamlDotNet.Serialization;
 using YamlDotNet.Serialization.NamingConventions;
+using Nox.Generator.Presentation.Api;
 
 namespace Nox.Generator;
 
@@ -86,9 +86,7 @@ public class NoxCodeGenerator : IIncrementalGenerator
                 {
                     ODataConfigurationGenerator.Generate(context, solutionNameSpace, solution);
 
-                    ODataApiGenerator.Generate(context, solutionNameSpace, solution);
-
-                    ApiControllerGenerator.Generate(context, solutionNameSpace, solution);
+                    ApiGenerator.Generate(context, solutionNameSpace, solution);
                 }
 
                 if (generate.Application)
@@ -97,14 +95,11 @@ public class NoxCodeGenerator : IIncrementalGenerator
                     ApplicationEventGenerator.Generate(context, solutionNameSpace, solution);
                 }
             }
-
-
         }
         catch (Exception e)
         {
             _errors.Add(e.Message);
         }
-
 
         if (_errors.Any())
         {
