+++ src/Nox.Generator/Presentation/Api/ApiGenerator.OData/ODataConfigurationGenerator.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.CodeAnalysis;
+using Nox.Generator.Common;
 using Nox.Solution;
 using System.Linq;
-using Nox.Generator.Common;
 
 namespace Nox.Generator;
 
@@ -23,7 +23,8 @@ internal static class ODataConfigurationGenerator
         code.AppendLine($"using Microsoft.AspNetCore.Http;");
         code.AppendLine($"using Microsoft.AspNetCore.OData;");
         code.AppendLine($"using Microsoft.OData.ModelBuilder;");
-        code.AppendLine($"using {solutionNameSpace}.Domain;"); 
+
+        code.AppendLine();
         code.AppendLine($"namespace {solutionNameSpace}.Presentation.Api.OData;");
         code.AppendLine();
 
@@ -31,40 +32,45 @@ internal static class ODataConfigurationGenerator
 
         // Class
         code.StartBlock();
-            // Method
-            code.AppendLine($"public static void Register(IServiceCollection services)");
-        
-            // Method content
-            code.StartBlock();
-                code.AppendLine($"ODataModelBuilder builder = new ODataConventionModelBuilder();");
-                code.AppendLine();
+        // Method
+        code.AppendLine($"public static void Register(IServiceCollection services)");
 
-                foreach (var entity in solution.Domain.Entities)
-                {
-                    context.CancellationToken.ThrowIfCancellationRequested();
-                    code.AppendLine($"builder.EntitySet<{entity.Name}>(\"{entity.PluralName}\");");
-                }
+        // Method content
+        code.StartBlock();
+        code.AppendLine($"ODataModelBuilder builder = new ODataConventionModelBuilder();");
+        code.AppendLine();
 
-                code.AppendLine();
-                code.AppendLine($"services.AddControllers()");
-                code.Indent();
-                    code.AppendLine($".AddOData(options => options");
-                    code.Indent();
-                        code.AppendLine($".Select()");
-                        code.AppendLine($".Filter()");
-                        code.AppendLine($".OrderBy()");
-                        code.AppendLine($".Count()");
-                        code.AppendLine($".Expand()");
-                        code.AppendLine($".SkipToken()");
-                        code.AppendLine($".SetMaxTop(100)");
-                        code.AppendLine($".AddRouteComponents(\"api\", builder.GetEdmModel())");
-                    code.UnIndent();
-                    code.AppendLine($");");
-                code.UnIndent();
-            // End method
-            code.EndBlock();
+        foreach (var entity in solution.Domain.Entities)
+        {
+            context.CancellationToken.ThrowIfCancellationRequested();
+            code.AppendLine($"builder.EntitySet<{entity.Name}>(\"{entity.PluralName}\");");
+        }
+
+        code.AppendLine();
+        code.AppendLine($"services.AddControllers()");
+        code.Indent();
+        code.AppendLine($".AddOData(options => options");
+        code.Indent();
+        code.AppendLine($".Select()");
+        code.AppendLine($".Filter()");
+        code.AppendLine($".OrderBy()");
+        code.AppendLine($".Count()");
+        code.AppendLine($".Expand()");
+        code.AppendLine($".SkipToken()");
+        code.AppendLine($".SetMaxTop(100)");
+        code.AppendLine($".AddRouteComponents(\"api\", builder.GetEdmModel())");
+        code.UnIndent();
+        code.AppendLine($");");
+        code.UnIndent();
+        // End method
+        code.EndBlock();
+
         // End class
         code.EndBlock();
+
+        ODataModelGenerator.Generate(context, solutionNameSpace, solution);
+
+        ODataDbContextGenerator.Generate(context, solutionNameSpace, solution);
 
         code.GenerateSourceCode();
     }
